Use a separate fd just for status reporting
If there is a way for attacker-controlled inputs to be quoted in output
messages in some way with newlines not sanitized (not impossible to
imagine), then using stdin or stdout for status messages may allow an
attacker to confuse the status parser into thinking an invalid
signature is actually valid.

Apparently I'm not the first to point this out:
https://bugs.python.org/issue18544

However, the clean solution (pass_fds, a no-CLOEXEC whitelist) only
exists in python3's subprocess module :(

A decent fallback IMO is having gpgv2 inherit an os.pipe (keeping
close_fds=False) and using that as the --status-fd.